+++ src/Serializer/AbstractItemNormalizer.php
@@ -130,7 +130,7 @@ abstract class AbstractItemNormalizer extends AbstractObjectNormalizer
     public function getSupportedTypes(?string $format): array
     {
         return [
-            'object' => true,
+            'object' => false,
         ];
     }
 
@@ -378,6 +378,8 @@ abstract class AbstractItemNormalizer extends AbstractObjectNormalizer
                     $params[] = $context[static::DEFAULT_CONSTRUCTOR_ARGUMENTS][$class][$key];
                 } elseif ($constructorParameter->isDefaultValueAvailable()) {
                     $params[] = $constructorParameter->getDefaultValue();
+                } elseif (!($context[self::REQUIRE_ALL_PROPERTIES] ?? $this->defaultContext[self::REQUIRE_ALL_PROPERTIES] ?? false) && $constructorParameter->hasType() && $constructorParameter->getType()->allowsNull()) {
+                    $params[] = null;
                 } else {
                     if (!isset($context['not_normalizable_value_exceptions'])) {
                         $missingConstructorArguments[] = $constructorParameter->name;